[feat]: Criação do CRUD completo de evento

+++ tech/src/main/java/chronos/tech/model/classes/PerfilAcesso.java
@@ -2,7 +2,7 @@ package chronos.tech.model.classes;
 
 public class PerfilAcesso {
 
-    private Integer id_perfil;
+    private Long id_perfil;
     private String nome_perfil;
     private String descricao;
